+++ packages/iov-bcp/types/connection.d.ts
@@ -172,8 +172,8 @@ export interface BlockchainConnection {
     readonly disconnect: () => void;
     readonly chainId: () => ChainId;
     readonly height: () => Promise<number>;
-    readonly getTicker: (ticker: TokenTicker) => Promise<Token | undefined>;
-    readonly getAllTickers: () => Promise<ReadonlyArray<Token>>;
+    readonly getToken: (ticker: TokenTicker) => Promise<Token | undefined>;
+    readonly getAllTokens: () => Promise<ReadonlyArray<Token>>;
     /**
      * Get the current account information (e.g. balance)
      *