+++ packages/iov-rise/src/riseconnection.ts
@@ -162,12 +162,12 @@ export class RiseConnection implements BlockchainConnection {
     };
   }
 
-  public async getTicker(searchTicker: TokenTicker): Promise<Token | undefined> {
-    const results = (await this.getAllTickers()).find(t => t.tokenTicker === searchTicker);
+  public async getToken(searchTicker: TokenTicker): Promise<Token | undefined> {
+    const results = (await this.getAllTokens()).find(t => t.tokenTicker === searchTicker);
     return results;
   }
 
-  public async getAllTickers(): Promise<ReadonlyArray<Token>> {
+  public async getAllTokens(): Promise<ReadonlyArray<Token>> {
     return [
       {
         tokenTicker: constants.primaryTokenTicker,